[added] Stage 1 of Structure Revamp - Harvesters/Generators [added] Stage 1 of Deed Revamp - Harvesters/Generators [disabled] Vehicle and Resource deeds are temporarily disabled until next commit. [required] Please apply SQL update. [added/modified] Various small changes.
NOTE: Server admins, please disregard the "Pickup" option that is present on all objects. It is part of a future patch.

/crossesfingers

git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1489 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/resource/ResourceHarvestType.h
@@ -10,14 +10,14 @@
 
 class ResourceHarvestType {
 public:
-	static const int SOLAR = 1;
-	static const int CHEMICAL = 2;
-	static const int FLORA = 3;
-	static const int GAS = 4;
-	static const int GEOTHERMAL = 5;
-	static const int MINERAL = 6;
-	static const int WATER = 7;
-	static const int WIND = 8;
-	static const int FUSION = 9;
+	static const uint8 SOLAR = 1;
+	static const uint8 CHEMICAL = 2;
+	static const uint8 FLORA = 3;
+	static const uint8 GAS = 4;
+	static const uint8 GEOTHERMAL = 5;
+	static const uint8 MINERAL = 6;
+	static const uint8 WATER = 7;
+	static const uint8 WIND = 8;
+	static const uint8 FUSION = 9;
 };
 #endif /* RESOURCEHARVESTTYPE_H_ */
